+++ packages/updating-element/CHANGELOG.md
@@ -22,4 +22,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - UpdatingElement moved from `lit-element` package to `updating-element` package.
-- Console warnings added for removed API and other element problems in developer mode. Set `static strictWarnings` to receive warnings for common mistakes; for example, when an update triggers another update.
+- Console warnings added for removed API and other element problems in developer mode. Some warnings are errors and are always issued while others are optional: making changes in update warns by default and can be toggled via `MyElement.disabledWarnings.add('change-in-update)`; special migration warnings are off by default and can be toggled via `MyElement.disabledWarnings.delete('migration')`.